+++ HISTORY.md
@@ -36,6 +36,7 @@
     * blocked xhr requests to from Snap! to s.b.e
     * the "message" reporter and watcher in the control category has been deprecated and moved to dev mode for backwards compatibility
     * removed old hidden "prefer smooth animations" setting (no longer used, old projects will continue to work just fine)
+    * removed old hidden "virtual keyboard" setting
 * **Notable Fixes:**
     * made scrollbars in the wardrobe and jukebox more responsive
     * fixed centering of menus, thanks, Brian Broll!
@@ -62,6 +63,7 @@
 * gui: made "Hyper blocks support" setting hidden in the gears menu
 * gui, objects, threads, store, translations: Removed old hidden "prefer smooth animations" setting
 * gui: removed experimental hidden "add scenes" option from the settings menus
+* gui, translations: removed old hidden "virtual keyboard" setting
 
 ### 2021-11-14
 * locale: contextualize translations